Address Luke's comments of ServerFactory related part.
