+++ runners/google-cloud-dataflow-java/worker/src/main/java/org/apache/beam/runners/dataflow/worker/DataflowRunnerHarness.java
@@ -106,7 +106,8 @@ public class DataflowRunnerHarness {
               ImmutableList.of(beamFnControlService, beamFnDataService, beamFnStateService),
               controlApiService);
 
-      loggingServer = serverFactory.create(beamFnLoggingService, loggingApiService);
+      loggingServer =
+          serverFactory.create(ImmutableList.of(beamFnLoggingService), loggingApiService);
 
       start(
           pipeline,
