+++ runners/java-fn-execution/src/main/java/org/apache/beam/runners/fnexecution/ServerFactory.java
@@ -20,6 +20,7 @@ package org.apache.beam.runners.fnexecution;
 import static com.google.common.base.Preconditions.checkArgument;
 
 import com.google.common.net.HostAndPort;
+import java.io.File;
 import java.io.IOException;
 import java.net.InetAddress;
 import java.net.InetSocketAddress;
@@ -30,30 +31,49 @@ import org.apache.beam.model.pipeline.v1.Endpoints;
 import org.apache.beam.sdk.fn.channel.SocketAddressFactory;
 import org.apache.beam.vendor.grpc.v1_13_1.io.grpc.BindableService;
 import org.apache.beam.vendor.grpc.v1_13_1.io.grpc.Server;
+import org.apache.beam.vendor.grpc.v1_13_1.io.grpc.ServerBuilder;
 import org.apache.beam.vendor.grpc.v1_13_1.io.grpc.ServerInterceptors;
 import org.apache.beam.vendor.grpc.v1_13_1.io.grpc.netty.NettyServerBuilder;
+import org.apache.beam.vendor.grpc.v1_13_1.io.netty.channel.epoll.EpollEventLoopGroup;
+import org.apache.beam.vendor.grpc.v1_13_1.io.netty.channel.epoll.EpollServerDomainSocketChannel;
+import org.apache.beam.vendor.grpc.v1_13_1.io.netty.channel.epoll.EpollServerSocketChannel;
+import org.apache.beam.vendor.grpc.v1_13_1.io.netty.channel.unix.DomainSocketAddress;
+import org.apache.beam.vendor.grpc.v1_13_1.io.netty.util.internal.ThreadLocalRandom;
 
 /** A {@link Server gRPC server} factory. */
 public abstract class ServerFactory {
-  /** Create a default {@link ServerFactory}. */
+  /** Create a default {@link InetSocketAddressServerFactory}. */
   public static ServerFactory createDefault() {
     return new InetSocketAddressServerFactory(UrlFactory.createDefault());
   }
 
-  /** Create a {@link ServerFactory} that uses the given url factory. */
+  /** Create a {@link InetSocketAddressServerFactory} that uses the given url factory. */
   public static ServerFactory createWithUrlFactory(UrlFactory urlFactory) {
     return new InetSocketAddressServerFactory(urlFactory);
   }
 
-  /** Create a {@link ServerFactory} that uses ports from a supplier. */
+  /** Create a {@link InetSocketAddressServerFactory} that uses ports from a supplier. */
   public static ServerFactory createWithPortSupplier(Supplier<Integer> portSupplier) {
     return new InetSocketAddressServerFactory(UrlFactory.createDefault(), portSupplier);
   }
 
-  /** Create a {@link ServerFactory} that uses the given url factory and ports from a supplier. */
+  /**
+   * Create a {@link InetSocketAddressServerFactory} that uses the given url factory and ports from
+   * a supplier.
+   */
   public static ServerFactory createWithUrlFactoryAndPortSupplier(
       UrlFactory urlFactory, Supplier<Integer> portSupplier) {
     return new InetSocketAddressServerFactory(urlFactory, portSupplier);
+  }
+
+  /** Create a {@link EpollSocket} */
+  public static ServerFactory createEpollSocket() {
+    return new EpollSocket();
+  }
+
+  /** Create a {@link EpollDomainSocket} */
+  public static ServerFactory createEpollDomainSocket() {
+    return new EpollDomainSocket();
   }
 
   /**
@@ -63,19 +83,12 @@ public abstract class ServerFactory {
    * GrpcContextHeaderAccessorProvider#interceptor()} to all incoming requests.
    */
   public abstract Server allocatePortAndCreate(
-      BindableService service, Endpoints.ApiServiceDescriptor.Builder builder) throws IOException;
-
-  /**
-   * Creates an instance of this server at the address specified by the given service descriptor.
-   * Server applies {@link GrpcContextHeaderAccessorProvider#interceptor()} to all incoming
-   * requests.
-   */
-  public abstract Server create(
-      BindableService service, Endpoints.ApiServiceDescriptor serviceDescriptor) throws IOException;
+      List<BindableService> services, Endpoints.ApiServiceDescriptor.Builder builder)
+      throws IOException;
 
   /**
    * Creates an instance of this server at the address specified by the given service descriptor and
-   * binded by multi services. Server applies {@link
+   * bound to multiple services. Server applies {@link
    * GrpcContextHeaderAccessorProvider#interceptor()} to all incoming requests.
    */
   public abstract Server create(
@@ -101,26 +114,13 @@ public abstract class ServerFactory {
 
     @Override
     public Server allocatePortAndCreate(
-        BindableService service, Endpoints.ApiServiceDescriptor.Builder apiServiceDescriptor)
+        List<BindableService> services, Endpoints.ApiServiceDescriptor.Builder apiServiceDescriptor)
         throws IOException {
       InetSocketAddress address =
           new InetSocketAddress(InetAddress.getLoopbackAddress(), portSupplier.get());
-      Server server = createServer(service, address);
+      Server server = createServer(services, address);
       apiServiceDescriptor.setUrl(urlFactory.createUrl(address.getHostName(), server.getPort()));
       return server;
-    }
-
-    @Override
-    public Server create(BindableService service, Endpoints.ApiServiceDescriptor serviceDescriptor)
-        throws IOException {
-      SocketAddress socketAddress = SocketAddressFactory.createFrom(serviceDescriptor.getUrl());
-      checkArgument(
-          socketAddress instanceof InetSocketAddress,
-          "%s %s requires a host:port socket address, got %s",
-          getClass().getSimpleName(),
-          ServerFactory.class.getSimpleName(),
-          serviceDescriptor.getUrl());
-      return createServer(service, (InetSocketAddress) socketAddress);
     }
 
     @Override
@@ -153,21 +153,108 @@ public abstract class ServerFactory {
                           service, GrpcContextHeaderAccessorProvider.interceptor())));
       return builder.build().start();
     }
+  }
 
-    private static Server createServer(BindableService service, InetSocketAddress socket)
+  /**
+   * Creates a {@link Server gRPC Server} using a Unix domain socket. Note that this requires <a
+   * href="http://netty.io/wiki/forked-tomcat-native.html">Netty TcNative</a> available to be able
+   * to provide a {@link EpollServerDomainSocketChannel}.
+   *
+   * <p>The unix domain socket is located at ${java.io.tmpdir}/fnapi${random[0-10000)}.sock
+   */
+  private static class EpollDomainSocket extends ServerFactory {
+    private static File getFileForPort(int port) {
+      return new File(System.getProperty("java.io.tmpdir"), String.format("fnapi%d.sock", port));
+    }
+
+    @Override
+    public Server allocatePortAndCreate(
+        List<BindableService> services, Endpoints.ApiServiceDescriptor.Builder apiServiceDescriptor)
         throws IOException {
-      // Note: Every ServerFactory should apply GrpcContextHeaderAccessorProvider to the service.
-      Server server =
-          NettyServerBuilder.forPort(socket.getPort())
-              .addService(
-                  ServerInterceptors.intercept(
-                      service, GrpcContextHeaderAccessorProvider.interceptor()))
-              // Set the message size to max value here. The actual size is governed by the
-              // buffer size in the layers above.
-              .maxMessageSize(Integer.MAX_VALUE)
-              .build();
-      server.start();
+      File tmp;
+      do {
+        tmp = getFileForPort(ThreadLocalRandom.current().nextInt(10000));
+      } while (tmp.exists());
+      apiServiceDescriptor.setUrl("unix://" + tmp.getAbsolutePath());
+      return create(services, apiServiceDescriptor.build());
+    }
+
+    @Override
+    public Server create(
+        List<BindableService> services, Endpoints.ApiServiceDescriptor serviceDescriptor)
+        throws IOException {
+      SocketAddress socketAddress = SocketAddressFactory.createFrom(serviceDescriptor.getUrl());
+      checkArgument(
+          socketAddress instanceof DomainSocketAddress,
+          "%s requires a Unix domain socket address, got %s",
+          EpollDomainSocket.class.getSimpleName(),
+          serviceDescriptor.getUrl());
+      return createServer(services, (DomainSocketAddress) socketAddress);
+    }
+
+    private static Server createServer(
+        List<BindableService> services, DomainSocketAddress domainSocket) throws IOException {
+      NettyServerBuilder builder =
+          NettyServerBuilder.forAddress(domainSocket)
+              .channelType(EpollServerDomainSocketChannel.class)
+              .workerEventLoopGroup(new EpollEventLoopGroup())
+              .bossEventLoopGroup(new EpollEventLoopGroup())
+              .maxMessageSize(Integer.MAX_VALUE);
+      for (BindableService service : services) {
+        // Wrap the service to extract headers
+        builder.addService(
+            ServerInterceptors.intercept(service, GrpcContextHeaderAccessorProvider.interceptor()));
+      }
+      return builder.build().start();
+    }
+  }
+
+  /**
+   * Creates a {@link Server gRPC Server} using an Epoll socket. Note that this requires <a
+   * href="http://netty.io/wiki/forked-tomcat-native.html">Netty TcNative</a> available to be able
+   * to provide a {@link EpollServerSocketChannel}.
+   *
+   * <p>The server is created listening any open port on "localhost".
+   */
+  private static class EpollSocket extends ServerFactory {
+    @Override
+    public Server allocatePortAndCreate(
+        List<BindableService> services, Endpoints.ApiServiceDescriptor.Builder apiServiceDescriptor)
+        throws IOException {
+      InetSocketAddress address = new InetSocketAddress(InetAddress.getLoopbackAddress(), 0);
+      Server server = createServer(services, address);
+      apiServiceDescriptor.setUrl(
+          HostAndPort.fromParts(address.getHostName(), server.getPort()).toString());
       return server;
+    }
+
+    @Override
+    public Server create(
+        List<BindableService> services, Endpoints.ApiServiceDescriptor serviceDescriptor)
+        throws IOException {
+      SocketAddress socketAddress = SocketAddressFactory.createFrom(serviceDescriptor.getUrl());
+      checkArgument(
+          socketAddress instanceof InetSocketAddress,
+          "%s requires a host:port socket address, got %s",
+          EpollSocket.class.getSimpleName(),
+          serviceDescriptor.getUrl());
+      return createServer(services, (InetSocketAddress) socketAddress);
+    }
+
+    private static Server createServer(List<BindableService> services, InetSocketAddress socket)
+        throws IOException {
+      ServerBuilder builder =
+          NettyServerBuilder.forAddress(socket)
+              .channelType(EpollServerSocketChannel.class)
+              .workerEventLoopGroup(new EpollEventLoopGroup())
+              .bossEventLoopGroup(new EpollEventLoopGroup())
+              .maxMessageSize(Integer.MAX_VALUE);
+      for (BindableService service : services) {
+        // Wrap the service to extract headers
+        builder.addService(
+            ServerInterceptors.intercept(service, GrpcContextHeaderAccessorProvider.interceptor()));
+      }
+      return builder.build().start();
     }
   }
